UPDATE: move player transformer to game
