+++ whist/core/game/hand.py
@@ -8,7 +8,6 @@ from whist.core.game.play_order import PlayOrder
 from whist.core.game.player_at_table import PlayerAtTable
 from whist.core.game.trick import Trick
 from whist.core.game.warnings import TrickNotDoneWarning
-from whist.core.user.player import Player
 
 
 class Hand:
@@ -36,10 +35,10 @@ class Hand:
         Returns the current trick.
         """
         if len(self._tricks) == 0:
-            self.deal()
+            self.deal(self._current_play_order)
         return self._tricks[-1]
 
-    def deal(self) -> Trick:
+    def deal(self, play_order: PlayOrder) -> Trick:
         """
         Deals the hand and starts the first trick.
         :return: the first trick
@@ -51,16 +50,16 @@ class Hand:
         deck = UnorderedCardContainer.full()
         card: Optional[Card] = None
         while deck:
-            player = self._current_play_order.next_player()
+            player = play_order.next_player()
             card = deck.pop_random()
             player.hand.add(card)
         self._trump = card.suit
 
-        first_trick = Trick(play_order=list(self._current_play_order), trump=self._trump)
+        first_trick = Trick(play_order=list(play_order), trump=self._trump)
         self._tricks.append(first_trick)
         return first_trick
 
-    def next_trick(self) -> Trick:
+    def next_trick(self, play_order: PlayOrder) -> Trick:
         """
         Starts the next trick.
         :return: the next trick
@@ -68,19 +67,11 @@ class Hand:
         """
         if not self._tricks[-1].done:
             raise TrickNotDoneWarning()
-        next_trick_order = self._winner_plays_first_card()
+        next_trick_order = self._winner_plays_first_card(play_order)
         next_trick = Trick(play_order=list(next_trick_order), trump=self._trump)
         self._tricks.append(next_trick)
         return next_trick
 
-    def get_player(self, player: Player) -> PlayerAtTable:
-        """
-        Retrieves the PlayerAtTable for the player given.
-        :param player: who needs it's counterpart at the table
-        :return: the player at table
-        """
-        return self._current_play_order.get_player(player)
-
-    def _winner_plays_first_card(self) -> PlayOrder:
+    def _winner_plays_first_card(self, play_order: PlayOrder) -> PlayOrder:
         winner: PlayerAtTable = self._tricks[-1].winner
-        return self._current_play_order.rotate(winner)
+        return play_order.rotate(winner)
